Move state of buyer's waiting for seller's account data to Phase 1
