+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/trade_state/TradePhaseBox.java
@@ -118,20 +118,20 @@ class TradePhaseBox {
                     switch (state) {
                         case INIT:
                             break;
+
                         case TAKER_SENT_TAKE_OFFER_REQUEST:
                         case MAKER_SENT_TAKE_OFFER_RESPONSE:
                         case TAKER_RECEIVED_TAKE_OFFER_RESPONSE:
+                        case SELLER_RECEIVED_BTC_ADDRESS:
+                        case BUYER_SENT_BTC_ADDRESS:
                             model.getPhaseIndex().set(0);
                             model.getRequestMediationButtonVisible().set(false);
                             model.getReportToMediatorButtonVisible().set(true);
-
-                        case SELLER_RECEIVED_BTC_ADDRESS:
-                        case BUYER_RECEIVED_ACCOUNT_DATA:
                             break;
 
                         case SELLER_SENT_ACCOUNT_DATA:
                         case SELLER_RECEIVED_FIAT_SENT_CONFIRMATION:
-                        case BUYER_SENT_BTC_ADDRESS:
+                        case BUYER_RECEIVED_ACCOUNT_DATA:
                         case BUYER_SENT_FIAT_SENT_CONFIRMATION:
                             model.getPhaseIndex().set(1);
                             model.getRequestMediationButtonVisible().set(false);
