+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/trade_state/states/BuyerState2b.java
@@ -19,16 +19,12 @@ package bisq.desktop.main.content.bisq_easy.open_trades.trade_state.states;
 
 import bisq.chat.bisqeasy.open_trades.BisqEasyOpenTradeChannel;
 import bisq.desktop.ServiceProvider;
-import bisq.desktop.components.controls.MaterialTextArea;
-import bisq.desktop.components.controls.MaterialTextField;
+import bisq.desktop.main.content.bisq_easy.components.WaitingAnimation;
+import bisq.desktop.main.content.bisq_easy.components.WaitingState;
 import bisq.desktop.components.controls.WrappingText;
-import bisq.desktop.components.overlay.Popup;
 import bisq.i18n.Res;
-import bisq.trade.TradeException;
 import bisq.trade.bisq_easy.BisqEasyTrade;
-import javafx.geometry.Insets;
-import javafx.scene.control.Button;
-import javafx.scene.layout.VBox;
+import javafx.scene.layout.HBox;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
@@ -68,67 +64,43 @@ public class BuyerState2b extends BaseState {
         public void onDeactivate() {
             super.onDeactivate();
         }
-
-        private void onConfirmFiatSent() {
-            sendSystemMessage(Res.get("bisqEasy.tradeState.info.buyer.phase2b.systemMessage", model.getQuoteCode()));
-            try {
-                bisqEasyTradeService.buyerConfirmFiatSent(model.getBisqEasyTrade());
-            } catch (TradeException e) {
-                new Popup().error(e).show();
-            }
-        }
     }
 
     @Getter
     private static class Model extends BaseState.Model {
-
         protected Model(BisqEasyTrade bisqEasyTrade, BisqEasyOpenTradeChannel channel) {
             super(bisqEasyTrade, channel);
         }
     }
 
     public static class View extends BaseState.View<Model, Controller> {
-        private final Button confirmFiatSentButton;
-        private final MaterialTextArea account;
-        private final MaterialTextField quoteAmount;
-        private final WrappingText headline;
+        private final WrappingText headline, info;
+        private final WaitingAnimation waitingAnimation;
 
         private View(Model model, Controller controller) {
             super(model, controller);
 
+            waitingAnimation = new WaitingAnimation(WaitingState.FIAT_PAYMENT_CONFIRMATION);
             headline = FormUtils.getHeadline();
-
-            quoteAmount = FormUtils.getTextField(Res.get("bisqEasy.tradeState.info.buyer.phase2b.quoteAmount"), "", false);
-            account = FormUtils.addTextArea(Res.get("bisqEasy.tradeState.info.buyer.phase2b.sellersAccount"), "", false);
-            account.setHelpText(Res.get("bisqEasy.tradeState.info.buyer.phase2b.reasonForPaymentInfo"));
-
-            confirmFiatSentButton = new Button();
-            confirmFiatSentButton.setDefaultButton(true);
-            VBox.setMargin(confirmFiatSentButton, new Insets(0, 0, 5, 0));
-
-            root.getChildren().addAll(
-                    headline,
-                    quoteAmount,
-                    account,
-                    confirmFiatSentButton);
+            info = FormUtils.getInfo();
+            HBox waitingInfo = createWaitingInfo(waitingAnimation, headline, info);
+            root.getChildren().add(waitingInfo);
         }
 
         @Override
         protected void onViewAttached() {
             super.onViewAttached();
 
-            headline.setText(Res.get("bisqEasy.tradeState.info.buyer.phase2b.headline", model.getFormattedQuoteAmount()));
-            quoteAmount.setText(model.getFormattedQuoteAmount());
-            account.setText(model.getBisqEasyTrade().getPaymentAccountData().get());
-            confirmFiatSentButton.setText(Res.get("bisqEasy.tradeState.info.buyer.phase2b.confirmFiatSent", model.getFormattedQuoteAmount()));
-            confirmFiatSentButton.setOnAction(e -> controller.onConfirmFiatSent());
+            headline.setText(Res.get("bisqEasy.tradeState.info.buyer.phase2b.headline"));
+            info.setText(Res.get("bisqEasy.tradeState.info.buyer.phase2b.info", model.getFormattedQuoteAmount()));
+            waitingAnimation.play();
         }
 
         @Override
         protected void onViewDetached() {
             super.onViewDetached();
 
-            confirmFiatSentButton.setOnAction(null);
+            waitingAnimation.stop();
         }
     }
 }
