+++ i18n/src/main/resources/bisq_easy.properties
@@ -544,26 +544,26 @@ bisqEasy.tradeState.info.phase5.exportTrade=Export trade data
 bisqEasy.tradeState.info.phase5.leaveChannel=Close trade
 
 # Trade State Info: Buyer
-bisqEasy.tradeState.info.buyer.phase1.btcAddress.headline=Fill in your Bitcoin address
-bisqEasy.tradeState.info.buyer.phase1.btcAddress=Bitcoin address
-bisqEasy.tradeState.info.buyer.phase1.btcAddress.prompt=Fill in your Bitcoin address
-bisqEasy.tradeState.info.buyer.phase1.btcAddress.help=If you have not set up a wallet yet, you can find help at the wallet guide
-bisqEasy.tradeState.info.buyer.phase1.walletHelpButton=Open wallet guide
-bisqEasy.tradeState.info.buyer.phase1.sendBtcAddress=Send Bitcoin address to the seller
-bisqEasy.tradeState.info.buyer.phase1.systemMessage=My bitcoin address is:\n{0}
+bisqEasy.tradeState.info.buyer.phase1a.btcAddress.headline=Fill in your Bitcoin address
+bisqEasy.tradeState.info.buyer.phase1a.btcAddress=Bitcoin address
+bisqEasy.tradeState.info.buyer.phase1a.btcAddress.prompt=Fill in your Bitcoin address
+bisqEasy.tradeState.info.buyer.phase1a.btcAddress.help=If you have not set up a wallet yet, you can find help at the wallet guide
+bisqEasy.tradeState.info.buyer.phase1a.walletHelpButton=Open wallet guide
+bisqEasy.tradeState.info.buyer.phase1a.sendBtcAddress=Send Bitcoin address to the seller
+bisqEasy.tradeState.info.buyer.phase1a.systemMessage=My bitcoin address is:\n{0}
 
-bisqEasy.tradeState.info.buyer.phase2a.headline=Wait for the seller's payment account data
-bisqEasy.tradeState.info.buyer.phase2a.info=You can use the chat below for getting in touch with the seller.
+bisqEasy.tradeState.info.buyer.phase1b.headline=Wait for the seller's payment account data
+bisqEasy.tradeState.info.buyer.phase1b.info=You can use the chat below for getting in touch with the seller.
 
-bisqEasy.tradeState.info.buyer.phase2b.headline=Send {0} to the seller''s payment account
-bisqEasy.tradeState.info.buyer.phase2b.quoteAmount=Amount to transfer
-bisqEasy.tradeState.info.buyer.phase2b.sellersAccount=Payment account of seller
-bisqEasy.tradeState.info.buyer.phase2b.reasonForPaymentInfo=Please leave the 'Reason for payment' field empty, in case you make a bank transfer
-bisqEasy.tradeState.info.buyer.phase2b.confirmFiatSent=Confirm payment of {0}
-bisqEasy.tradeState.info.buyer.phase2b.systemMessage=I initiated the {0} payment.
+bisqEasy.tradeState.info.buyer.phase2a.headline=Send {0} to the seller''s payment account
+bisqEasy.tradeState.info.buyer.phase2a.quoteAmount=Amount to transfer
+bisqEasy.tradeState.info.buyer.phase2a.sellersAccount=Payment account of seller
+bisqEasy.tradeState.info.buyer.phase2a.reasonForPaymentInfo=Please leave the 'Reason for payment' field empty, in case you make a bank transfer
+bisqEasy.tradeState.info.buyer.phase2a.confirmFiatSent=Confirm payment of {0}
+bisqEasy.tradeState.info.buyer.phase2a.systemMessage=I initiated the {0} payment.
 
-bisqEasy.tradeState.info.buyer.phase2c.headline=Wait for the seller to confirm receipt of payment
-bisqEasy.tradeState.info.buyer.phase2c.info=Once the seller has received your payment of {0}, they will start the Bitcoin payment to your provided Bitcoin address.
+bisqEasy.tradeState.info.buyer.phase2b.headline=Wait for the seller to confirm receipt of payment
+bisqEasy.tradeState.info.buyer.phase2b.info=Once the seller has received your payment of {0}, they will start the Bitcoin payment to your provided Bitcoin address.
 
 bisqEasy.tradeState.info.buyer.phase3a.headline=Wait for the seller's Bitcoin payment
 bisqEasy.tradeState.info.buyer.phase3a.info=The seller need to start the Bitcoin payment to your provided Bitcoin address.
